Move storage coupled processors
