+++ snuba/datasets/plans/query_plan.py
@@ -7,9 +7,9 @@ from typing import Callable, Sequence
 from snuba.clickhouse.processors import QueryProcessor
 from snuba.clickhouse.query import Query
 from snuba.clickhouse.sql import SqlQuery
+from snuba.reader import Reader
 from snuba.request import Request
 from snuba.request.request_settings import RequestSettings
-from snuba.reader import Reader
 from snuba.web import QueryResult
 
 QueryRunner = Callable[[Query, RequestSettings, Reader[SqlQuery]], QueryResult]
